Rename log_parse.py -> mdb_log_parse.py
